ipq807x: update current partition directly
Right now we use A/B partition during the upgrade, in order to avoid
any mistake causing the device get bricked. But the vendor partition
is too strange so we need to adjust the partition for more space to
use and this obviously breaks the A/B partition.

However it is fine to upgrade directly, then just simply modify the
the code to match this behavior.

Signed-off-by: Tianling Shen <cnsztl@immortalwrt.org>
Signed-off-by: YiZhen Choo <yizhen.c02@gmail.com>